more chat tweaks
- moved admin controls to popup menu on gear icon instead of inside the members fragment
- fixed the bug where 'X' would appear on rooms you were still part of
