+++ GroupChat/app/src/main/res/values-night/themes.xml
@@ -30,7 +30,9 @@
         <item name="colorOnSurface">@color/white</item>
         <item name="colorTinted">@color/grey_1000</item>
         <item name="colorOnTinted">@color/grey_200</item>
+        <item name="colorPopupBackground">@color/grey_900</item>
         <item name="android:spinnerDropDownItemStyle">@style/SpinnerItem</item>
+        <item name="android:popupMenuStyle">@style/PopupMenu</item>
     </style>
     <!-- Indigo/Green application theme. -->
     <style name="Theme.IndigoGreen" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
@@ -63,7 +65,9 @@
         <item name="colorOnSurface">@color/white</item>
         <item name="colorTinted">@color/grey_1000</item>
         <item name="colorOnTinted">@color/grey_200</item>
+        <item name="colorPopupBackground">@color/grey_900</item>
         <item name="android:spinnerDropDownItemStyle">@style/SpinnerItem</item>
+        <item name="android:popupMenuStyle">@style/PopupMenu</item>
     </style>
     <!-- Blue Grey/Orange application theme. -->
     <style name="Theme.GreyOrange" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
@@ -96,7 +100,9 @@
         <item name="colorOnSurface">@color/white</item>
         <item name="colorTinted">@color/grey_1000</item>
         <item name="colorOnTinted">@color/grey_200</item>
+        <item name="colorPopupBackground">@color/grey_900</item>
         <item name="android:spinnerDropDownItemStyle">@style/SpinnerItem</item>
+        <item name="android:popupMenuStyle">@style/PopupMenu</item>
     </style>
     <!-- Tab Text Styling -->
     <style name="TabText" parent="TextAppearance.Design.Tab">
@@ -109,5 +115,9 @@
         <item name="android:background">?attr/colorSurface</item>
         <item name="android:textColor">?attr/colorOnSurface</item>
         <item name="android:textAppearance">?android:attr/textAppearanceSmall</item>
+    </style>
+    <!-- PopUp Menu Styling -->
+    <style name="PopupMenu" parent="@android:style/Widget.PopupMenu">
+        <item name="android:popupBackground">?attr/colorPopupBackground</item>
     </style>
 </resources>
